+++ Models/Square.cs
@@ -8,9 +8,9 @@ public sealed class Square : Shape
 {
     private Point _point1, _point2, _point3, _point4;
     private static float InnerR => R / (float)Math.Sqrt(2);
-    public Square(int x, int y) : base(x, y) { }
+    public Square(double x, double y) : base(x, y) { }
 
-    public override bool IsInside(int newX, int newY)
+    public override bool IsInside(double newX, double newY)
     {
         return x - InnerR <= newX && newX <= x + InnerR && y - InnerR <= newY && newY <= y + InnerR;
     }
